backtester/requirements.txt
requests==2.31.0
pandas==2.2.0
psycopg2-binary
python-dotenv==1.0.1
